Extract utilities from AbstractAntUITest into AntUITestUtil
In order to remove the inheritance hierarchy of Ant UI tests for JUnit 5
migration, this moves the utility methods embedded into the inheritance
hierarchy via the AbstractAntUITest root class to the dedicated
AntUITestUtil.
