+++ ant/org.eclipse.ant.tests.ui/Ant Editor Tests/org/eclipse/ant/tests/ui/editor/OccurrencesFinderTests.java
@@ -13,6 +13,7 @@
  *******************************************************************************/
 package org.eclipse.ant.tests.ui.editor;
 
+import static org.eclipse.ant.tests.ui.testplugin.AntUITestUtil.getIFile;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
