+++ ant/org.eclipse.ant.tests.ui/test plugin/org/eclipse/ant/tests/ui/testplugin/AntUITestUtil.java
@@ -10,8 +10,10 @@
  *******************************************************************************/
 package org.eclipse.ant.tests.ui.testplugin;
 
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
@@ -19,6 +21,7 @@ import java.net.URL;
 import org.eclipse.ant.internal.ui.AntUIPlugin;
 import org.eclipse.ant.internal.ui.IAntUIPreferenceConstants;
 import org.eclipse.ant.tests.ui.debug.TestAgainException;
+import org.eclipse.core.externaltools.internal.IExternalToolConstants;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IFolder;
 import org.eclipse.core.resources.IProject;
@@ -30,10 +33,22 @@ import org.eclipse.debug.core.DebugEvent;
 import org.eclipse.debug.core.DebugPlugin;
 import org.eclipse.debug.core.ILaunch;
 import org.eclipse.debug.core.ILaunchConfiguration;
+import org.eclipse.debug.core.ILaunchConfigurationWorkingCopy;
 import org.eclipse.debug.core.ILaunchManager;
 import org.eclipse.debug.core.model.IProcess;
 import org.eclipse.jdt.core.IJavaProject;
 import org.eclipse.jdt.core.JavaCore;
+import org.eclipse.jface.text.BadLocationException;
+import org.eclipse.jface.text.BadPositionCategoryException;
+import org.eclipse.jface.text.IDocument;
+import org.eclipse.jface.text.IDocumentPartitioner;
+import org.eclipse.jface.text.ITypedRegion;
+import org.eclipse.jface.text.Position;
+import org.eclipse.swt.graphics.Color;
+import org.eclipse.swt.widgets.Display;
+import org.eclipse.ui.console.IHyperlink;
+import org.eclipse.ui.internal.console.ConsoleHyperlinkPosition;
+import org.eclipse.ui.internal.console.IOConsolePartition;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.FrameworkUtil;
 
@@ -192,5 +207,132 @@ public final class AntUITestUtil {
 	 */
 	public static IProject getProject() {
 		return ResourcesPlugin.getWorkspace().getRoot().getProject(ProjectHelper.PROJECT_NAME);
+	}
+
+	/**
+	 * Launches the Ant build with the build file name (no extension).
+	 *
+	 * @param buildFileName
+	 *            the build file
+	 * @param arguments
+	 *            the ant arguments
+	 */
+	public static void launch(String buildFileName, String arguments) throws CoreException {
+		ILaunchConfiguration config = getLaunchConfiguration(buildFileName);
+		assertNotNull("Could not locate launch configuration for " + buildFileName, config); //$NON-NLS-1$
+		ILaunchConfigurationWorkingCopy copy = config.getWorkingCopy();
+		copy.setAttribute(IExternalToolConstants.ATTR_TOOL_ARGUMENTS, arguments);
+		launchAndTerminate(copy, 20000);
+	}
+
+	/**
+	 * Launches the Ant build with the build file name (no extension).
+	 *
+	 * @param buildFileName
+	 *            the ant build file name
+	 */
+	public static void launch(String buildFileName) throws CoreException {
+		ILaunchConfiguration config = getLaunchConfiguration(buildFileName);
+		assertNotNull("Could not locate launch configuration for " + buildFileName, config); //$NON-NLS-1$
+		launchAndTerminate(config, 20000);
+	}
+
+	/**
+	 * Returns the {@link IFile} for the given build file name
+	 *
+	 * @return the associated {@link IFile} for the given build file name
+	 */
+	public static IFile getIFile(String buildFileName) {
+		return getProject().getFolder("buildfiles").getFile(buildFileName); //$NON-NLS-1$
+	}
+
+	/**
+	 * Returns the {@link File} for the given build file name
+	 *
+	 * @return the {@link File} for the given build file name
+	 */
+	public static File getBuildFile(String buildFileName) {
+		IFile file = getIFile(buildFileName);
+		assertTrue("Could not find build file named: " + buildFileName, file.exists()); //$NON-NLS-1$
+		return file.getLocation().toFile();
+	}
+
+	/**
+	 * Returns the contents of the given {@link BufferedReader} as a {@link String}
+	 *
+	 * @return the contents of the given {@link BufferedReader} as a {@link String}
+	 */
+	public static String getReaderContentAsString(BufferedReader bufferedReader) {
+		StringBuilder result = new StringBuilder();
+		try {
+			String line = bufferedReader.readLine();
+
+			while (line != null) {
+				if (result.length() != 0) {
+					result.append(System.lineSeparator());
+				}
+				result.append(line);
+				line = bufferedReader.readLine();
+			}
+		}
+		catch (IOException e) {
+			AntUIPlugin.log(e);
+			return null;
+		}
+
+		return result.toString();
+	}
+
+	/**
+	 * Returns the {@link Color} at the given offset on the given document, or
+	 * <code>null</code> if there is no {@link Color} at that offset on the
+	 * document.
+	 *
+	 * @return the {@link Color} at the given offset on the given document or
+	 *         <code>null</code>
+	 */
+	public static Color getColorAtOffset(int offset, IDocument document) throws BadLocationException {
+		if (document != null) {
+			IDocumentPartitioner partitioner = document.getDocumentPartitioner();
+			if (partitioner != null) {
+				ITypedRegion[] regions = partitioner.computePartitioning(offset,
+						document.getLineInformationOfOffset(offset).getLength());
+				if (regions.length > 0) {
+					IOConsolePartition partition = (IOConsolePartition) regions[0];
+					return partition.getColor();
+				}
+			}
+		}
+		return null;
+	}
+
+	/**
+	 * Returns the {@link IHyperlink} at the given offset on the given document, or
+	 * <code>null</code> if there is no {@link IHyperlink} at that offset on the
+	 * document.
+	 *
+	 * @return the {@link IHyperlink} at the given offset on the given document or
+	 *         <code>null</code>
+	 */
+	public static IHyperlink getHyperlink(int offset, IDocument doc) {
+		if (offset >= 0 && doc != null) {
+			Position[] positions = null;
+			try {
+				positions = doc.getPositions(ConsoleHyperlinkPosition.HYPER_LINK_CATEGORY);
+			} catch (BadPositionCategoryException ex) {
+				// no links have been added
+				return null;
+			}
+			for (Position position : positions) {
+				if (offset >= position.getOffset() && offset <= (position.getOffset() + position.getLength())) {
+					return ((ConsoleHyperlinkPosition) position).getHyperLink();
+				}
+			}
+		}
+		return null;
+	}
+
+	public static void activateLink(final IHyperlink link) {
+		Display.getDefault().asyncExec(() -> link.linkActivated());
 	}
 }
